chore: Contract's code reorganization ; added library feature for reusing contract
