+++ .cargo/config
@@ -1,6 +1,4 @@
 [alias]
 wasm = "build --release --target wasm32-unknown-unknown --lib"
 wasm-debug = "build --target wasm32-unknown-unknown --lib"
-unit-test = "test --lib"
-integration-test = "test --test integration"
 schema = "run --bin schema"
