+++ Cargo.toml
@@ -4,8 +4,14 @@ authors = ["{{authors}}"]
 version = "0.1.0"
 edition = "2021"
 
+
+[features]
+library = []
+
+
 [lib]
 crate-type = ["cdylib", "rlib"]
+
 
 [dependencies]
 cosmwasm-std = { version = "1", features = ["staking"] }
