+++ src/multitest/mod.rs
@@ -1,0 +1,4 @@
+
+
+// TODO tkulik: write simple multitest
+
